+++ app/handlers/apiv1/post.go
@@ -2,6 +2,7 @@ package apiv1
 
 import (
 	"github.com/getfider/fider/app/actions"
+	"github.com/getfider/fider/app/models"
 	"github.com/getfider/fider/app/pkg/web"
 	"github.com/getfider/fider/app/tasks"
 )
@@ -58,6 +59,35 @@ func UpdatePost() web.HandlerFunc {
 		if err != nil {
 			return c.Failure(err)
 		}
+
+		return c.Ok(web.Map{})
+	}
+}
+
+// SetResponse changes current post staff response
+func SetResponse() web.HandlerFunc {
+	return func(c web.Context) error {
+		input := new(actions.SetResponse)
+		if result := c.BindTo(input); !result.Ok {
+			return c.HandleValidation(result)
+		}
+
+		post, err := c.Services().Posts.GetByNumber(input.Model.Number)
+		if err != nil {
+			return c.Failure(err)
+		}
+
+		prevStatus := post.Status
+		if input.Model.Status == models.PostDuplicate {
+			err = c.Services().Posts.MarkAsDuplicate(post, input.Original)
+		} else {
+			err = c.Services().Posts.SetResponse(post, input.Model.Text, input.Model.Status)
+		}
+		if err != nil {
+			return c.Failure(err)
+		}
+
+		c.Enqueue(tasks.NotifyAboutStatusChange(post, prevStatus))
 
 		return c.Ok(web.Map{})
 	}
